pyproject.toml
[tool.poetry]
name = "bpl"
version = "0.1.0"
description = "Минимальная доска объявлений на Django с кастомным пользователем и REST API"
authors = ["ANTAQ <mymillions@ya.ru>"]
readme = "README.md"
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
Django = ">=5.2.7,<6.0.0"
djangorestframework = ">=3.16.1,<4.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
python-dotenv = ">=1.1.1,<2.0.0"
gunicorn = ">=23.0.0,<24.0.0"
celery = ">=5.4.0,<6.0"
redis = ">=5.0,<6.0"
djangorestframework-simplejwt = "^5.5.1"

[tool.poetry.dev-dependencies]
flake8 = ">=7.3.0,<8.0.0"
pytest = ">=8.4.2,<9.0.0"
pytest-django = ">=4.11.1,<5.0.0"
coverage = ">=7.10.7,<8.0.0"
ipython = ">=9.6.0,<10.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
